backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
python-dotenv==1.0.0
PyJWT==2.8.0
django-simpleui
requests==2.31.0
Pillow==9.5.0
mysqlclient==2.0.0
django-filter
faker==19.13.0